feat: add python test for dtype
Previously, there was an issue in case we passed dtypes that weren't the
same as the Python FFI function we define in Rust. MDAV is generic, so
we could use f32 or f64. In this package, though, we use f64 in our Rust
wrapper. This leads to only numpy arrays with float64 dtype working.
However, we also want float32, so we wrapped it in a python function
test that it works either way.