fix: move boto3 client to global variable
Fixes bizarre fastapi/uvicorn error. boto3.client cannot be instantiated
for the first time on EC2 in a dependency injected function. Has nothing
to do with an LRU cache or anything else. Just can't be instantiated in
a DI function, for some reason. Has something to do with fetching the
IAM roles, etc. Unclear, but this fixes it without reducing testability.